[Angular] refactoring Jwt
