+++ src/main/java/tech/jhipster/lite/generator/client/angular/core/application/AngularApplicationService.java
@@ -16,8 +16,4 @@ public class AngularApplicationService {
   public void addAngular(Project project) {
     angularService.addAngular(project);
   }
-
-  public void addJwtAngular(Project project) {
-    angularService.addJwtAngular(project);
-  }
 }
