+++ src/main/java/tech/jhipster/lite/generator/client/angular/core/domain/AngularService.java
@@ -4,5 +4,4 @@ import tech.jhipster.lite.generator.project.domain.Project;
 
 public interface AngularService {
   void addAngular(Project project);
-  void addJwtAngular(Project project);
 }
