+++ src/main/java/tech/jhipster/lite/generator/client/angular/core/infrastructure/primary/rest/AngularResource.java
@@ -32,13 +32,4 @@ class AngularResource {
     Project project = ProjectDTO.toProject(projectDTO);
     angularApplicationService.addAngular(project);
   }
-
-  @Operation(summary = "Add Angular with authentication JWT")
-  @ApiResponse(responseCode = "500", description = "An error occurred while adding Angular with authentication JWT")
-  @PostMapping("/jwt")
-  @GeneratorStep(id = "angular-jwt")
-  public void addJwtAngular(@RequestBody ProjectDTO projectDTO) {
-    Project project = ProjectDTO.toProject(projectDTO);
-    angularApplicationService.addJwtAngular(project);
-  }
 }
