+++ src/main/java/tech/jhipster/lite/generator/client/angular/security/jwt/domain/AngularJwtService.java
@@ -1,0 +1,7 @@
+package tech.jhipster.lite.generator.client.angular.security.jwt.domain;
+
+import tech.jhipster.lite.generator.project.domain.Project;
+
+public interface AngularJwtService {
+  void addJwtAngular(Project project);
+}
