+++ src/main/java/tech/jhipster/lite/generator/project/domain/GeneratorAction.java
@@ -48,6 +48,7 @@ public class GeneratorAction {
 
   public static final String ANGULAR = "angular";
   public static final String ANGULAR_STYLED = "angular-styled";
+  public static final String ANGULAR_JWT = "angular-jwt";
 
   public static final String LIQUIBASE = "liquibase";
   public static final String LIQUIBASE_USER_AND_AUTHORITY_CHANGELOGS = "liquibase-user-and-authority-changelogs";
