+++ src/test/java/tech/jhipster/lite/generator/client/angular/core/application/AngularApplicationServiceIT.java
@@ -2,7 +2,6 @@ package tech.jhipster.lite.generator.client.angular.core.application;
 
 import static tech.jhipster.lite.TestUtils.tmpProjectWithPackageJsonComplete;
 import static tech.jhipster.lite.generator.client.angular.core.application.AngularAssert.assertAngularFiles;
-import static tech.jhipster.lite.generator.client.angular.core.application.AngularAssert.assertAppJwt;
 import static tech.jhipster.lite.generator.client.angular.core.application.AngularAssert.assertConfigFiles;
 import static tech.jhipster.lite.generator.client.angular.core.application.AngularAssert.assertDependencies;
 import static tech.jhipster.lite.generator.client.angular.core.application.AngularAssert.assertDevDependencies;
@@ -30,20 +29,5 @@ class AngularApplicationServiceIT {
     assertScripts(project);
     assertConfigFiles(project);
     assertAngularFiles(project);
-  }
-
-  @Test
-  void shouldAddJwtAngular() {
-    Project project = tmpProjectWithPackageJsonComplete();
-
-    angularApplicationService.addAngular(project);
-    angularApplicationService.addJwtAngular(project);
-
-    assertDevDependencies(project);
-    assertDependencies(project);
-    assertScripts(project);
-    assertConfigFiles(project);
-    assertAngularFiles(project);
-    assertAppJwt(project);
   }
 }
