+++ src/test/java/tech/jhipster/lite/generator/client/angular/core/domain/AngularDomainServiceTest.java
@@ -50,21 +50,6 @@ class AngularDomainServiceTest {
   }
 
   @Test
-  void shouldAddJwtAngular() {
-    Project project = tmpProjectWithPackageJson();
-    when(npmService.getVersion(anyString(), anyString())).thenReturn(Optional.of("0.0.0"));
-
-    assertThatCode(() -> angularDomainService.addJwtAngular(project)).doesNotThrowAnyException();
-  }
-
-  @Test
-  void shouldNotAddJwtAngular() {
-    Project project = tmpProjectWithPackageJson();
-
-    assertThatThrownBy(() -> angularDomainService.addJwtAngular(project)).isExactlyInstanceOf(GeneratorException.class);
-  }
-
-  @Test
   void shouldAddDependencies() {
     Project project = tmpProjectWithPackageJson();
     when(npmService.getVersion(anyString(), anyString())).thenReturn(Optional.of("0.0.0"));
@@ -99,23 +84,6 @@ class AngularDomainServiceTest {
   }
 
   @Test
-  void shouldAddJwtDependencies() {
-    Project project = tmpProjectWithPackageJson();
-    when(npmService.getVersion(anyString(), anyString())).thenReturn(Optional.of("0.0.0"));
-
-    angularDomainService.addJwtDependencies(project);
-
-    verify(npmService, times(1)).addDependency(any(Project.class), anyString(), anyString());
-  }
-
-  @Test
-  void shouldNotAddJwtDependencies() {
-    Project project = tmpProjectWithPackageJson();
-
-    assertThatThrownBy(() -> angularDomainService.addJwtDependencies(project)).isExactlyInstanceOf(GeneratorException.class);
-  }
-
-  @Test
   void shouldAddScripts() {
     Project project = tmpProjectWithPackageJson();
 
@@ -134,29 +102,11 @@ class AngularDomainServiceTest {
   }
 
   @Test
-  void shouldAddJwtFiles() {
-    Project project = tmpProjectWithPackageJson();
-
-    angularDomainService.addJwtFiles(project);
-
-    verify(projectRepository, times(1)).template(any(Project.class), anyString(), anyString(), anyString());
-  }
-
-  @Test
   void shouldAddAngularFiles() {
     Project project = tmpProjectWithPackageJson();
 
     angularDomainService.addAngularFiles(project);
 
     verify(projectRepository, times(15)).template(any(Project.class), anyString(), anyString(), anyString());
-  }
-
-  @Test
-  void shouldAddAngularJwtFiles() {
-    Project project = tmpProjectWithPackageJson();
-
-    angularDomainService.addAngularJwtFiles(project);
-
-    verify(projectRepository, times(9)).template(any(Project.class), anyString(), anyString(), anyString());
   }
 }
