+++ src/test/java/tech/jhipster/lite/generator/client/angular/security/jwt/infrastructure/primary/rest/AngularJwtResourceIT.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.client.angular.core.infrastructure.primary.rest;
+package tech.jhipster.lite.generator.client.angular.security.jwt.infrastructure.primary.rest;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -12,14 +12,14 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import tech.jhipster.lite.IntegrationTest;
 import tech.jhipster.lite.TestUtils;
-import tech.jhipster.lite.generator.client.angular.core.application.AngularAssert;
+import tech.jhipster.lite.generator.client.angular.security.jwt.application.AngularJwtAssert;
 import tech.jhipster.lite.generator.init.application.InitApplicationService;
 import tech.jhipster.lite.generator.project.domain.Project;
 import tech.jhipster.lite.generator.project.infrastructure.primary.dto.ProjectDTO;
 
 @IntegrationTest
 @AutoConfigureMockMvc
-class AngularResourceIT {
+class AngularJwtResourceIT {
 
   @Autowired
   MockMvc mockMvc;
@@ -28,7 +28,7 @@ class AngularResourceIT {
   InitApplicationService initApplicationService;
 
   @Test
-  void shouldAddAngular() throws Exception {
+  void shouldAddJwtAngular() throws Exception {
     ProjectDTO projectDTO = readFileToObject("json/chips.json", ProjectDTO.class).folder(tmpDirForTest());
     Project project = ProjectDTO.toProject(projectDTO);
     initApplicationService.init(project);
@@ -36,9 +36,12 @@ class AngularResourceIT {
     mockMvc
       .perform(post("/api/clients/angular").contentType(MediaType.APPLICATION_JSON).content(TestUtils.convertObjectToJsonBytes(projectDTO)))
       .andExpect(status().isOk());
+    mockMvc
+      .perform(
+        post("/api/clients/angular/jwt").contentType(MediaType.APPLICATION_JSON).content(TestUtils.convertObjectToJsonBytes(projectDTO))
+      )
+      .andExpect(status().isOk());
 
-    AngularAssert.assertDevDependencies(project);
-    AngularAssert.assertScripts(project);
-    AngularAssert.assertConfigFiles(project);
+    AngularJwtAssert.assertAppJwt(project);
   }
 }
